+++ core/src/main/java/io/temporal/samples/nexus_messaging/caller/CallerWorkflow.java
@@ -1,0 +1,11 @@
+package io.temporal.samples.nexus_messaging.caller;
+
+import io.temporal.workflow.WorkflowInterface;
+import io.temporal.workflow.WorkflowMethod;
+import java.util.List;
+
+@WorkflowInterface
+public interface CallerWorkflow {
+  @WorkflowMethod
+  List<String> run();
+}
