+++ core/src/main/java/io/temporal/samples/nexus_messaging/caller_remote/CallerRemoteWorkflow.java
@@ -1,0 +1,11 @@
+package io.temporal.samples.nexus_messaging.caller_remote;
+
+import io.temporal.workflow.WorkflowInterface;
+import io.temporal.workflow.WorkflowMethod;
+import java.util.List;
+
+@WorkflowInterface
+public interface CallerRemoteWorkflow {
+  @WorkflowMethod
+  List<String> run();
+}
