+++ core/src/main/java/io/temporal/samples/nexus_messaging/handler/GreetingActivity.java
@@ -1,0 +1,12 @@
+package io.temporal.samples.nexus_messaging.handler;
+
+import io.temporal.activity.ActivityInterface;
+import io.temporal.activity.ActivityMethod;
+import io.temporal.samples.nexus_messaging.service.Language;
+
+@ActivityInterface
+public interface GreetingActivity {
+  // Simulates a call to a remote greeting service. Returns null if the language is not supported.
+  @ActivityMethod
+  String callGreetingService(Language language);
+}
