+++ core/src/main/java/io/temporal/samples/nexus_messaging/service/Language.java
@@ -1,0 +1,11 @@
+package io.temporal.samples.nexus_messaging.service;
+
+public enum Language {
+  ARABIC,
+  CHINESE,
+  ENGLISH,
+  FRENCH,
+  HINDI,
+  PORTUGUESE,
+  SPANISH
+}
