--- core/src/main/java/io/temporal/samples/nexus_sync_operations/service/Language.java
@@ -1,11 +1,0 @@
-package io.temporal.samples.nexus_sync_operations.service;
-
-public enum Language {
-  ARABIC,
-  CHINESE,
-  ENGLISH,
-  FRENCH,
-  HINDI,
-  PORTUGUESE,
-  SPANISH
-}
