--- core/src/main/java/io/temporal/samples/nexusmessaging/caller/MessageCallerWorkflow.java
@@ -1,10 +1,0 @@
-package io.temporal.samples.nexusmessaging.caller;
-
-import io.temporal.workflow.WorkflowInterface;
-import io.temporal.workflow.WorkflowMethod;
-
-@WorkflowInterface
-public interface MessageCallerWorkflow {
-  @WorkflowMethod
-  String sentMessage();
-}
