--- core/src/main/java/io/temporal/samples/nexusmessaging/service/description.md
@@ -1,6 +1,0 @@
-## Service: [SampleNexusService](https://github.com/temporalio/samples-java/blob/main/core/src/main/java/io/temporal/samples/nexusmessaging/service/SampleNexusService.java)
- - operation: `signalWorkflow`
- - operation: `queryWorkflow`
- - operation: `updateWorkflow`
-
-See https://github.com/temporalio/samples-java/blob/main/core/src/main/java/io/temporal/samples/nexusmessaging/service/SampleNexusService.java for Input / Output types.
